dtsrs_app/requirements.txt
blinker==1.8.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
firebase-admin==6.5.0
Flask==3.0.3
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
msgpack==1.0.8
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
pytz==2024.1
requests==2.31.0
rsa==4.9
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
wxPython==4.2.1
